+++ spec/features/visitors/anyone_can_complete_a_journey_spec.rb
@@ -187,6 +187,35 @@ feature "Anyone can start a journey" do
             .to eql("A second piece of further information")
         end
       end
+
+      context "when extended question is of type single" do
+        scenario "a single text field is displayed" do
+          start_journey_from_category_and_go_to_question(category: "extended-checkboxes-question.json")
+
+          check("Yes")
+
+          expect(page).to have_selector("input#answer-yes-further-information-field")
+        end
+      end
+
+      context "when extended question is of type long" do
+        scenario "a long text area is displayed" do
+          start_journey_from_category_and_go_to_question(category: "extended-long-answer-checkboxes-question.json")
+
+          check("Yes")
+
+          expect(page).to have_selector("textarea#answer-yes-further-information-field")
+        end
+      end
+
+      context "when there is no extended question" do
+        scenario "no extra text field is displayed" do
+          start_journey_from_category_and_go_to_question(category: "checkboxes-question.json")
+
+          expect(page).to_not have_selector("textarea#answer-yes-further-information-field")
+          expect(page).to_not have_selector("input#answer-yes-further-information-field")
+        end
+      end
     end
 
     context "when Contentful entry is of type radios" do
@@ -204,6 +233,35 @@ feature "Anyone can start a journey" do
           expect(page).to have_checked_field("Catering")
           expect(find_field("answer-further-information-field").value)
             .to eql("The school needs the kitchen cleaned once a day")
+        end
+      end
+
+      context "when extended question is of type single" do
+        scenario "a single text field is displayed" do
+          start_journey_from_category_and_go_to_question(category: "extended-radio-question.json")
+
+          choose("Catering")
+
+          expect(page).to have_selector("input#answer-further-information-field")
+        end
+      end
+
+      context "when extended question is of type long" do
+        scenario "a long text area is displayed" do
+          start_journey_from_category_and_go_to_question(category: "extended-long-answer-radio-question.json")
+
+          choose("Catering")
+
+          expect(page).to have_selector("textarea#answer-further-information-field")
+        end
+      end
+
+      context "when there is no extended question" do
+        scenario "no extra text field is displayed" do
+          start_journey_from_category_and_go_to_question(category: "radio-question.json")
+
+          expect(page).to_not have_selector("textarea#answer-further-information-field")
+          expect(page).to_not have_selector("input#answer-further-information-field")
         end
       end
     end