+++ CHANGELOG.md
@@ -29,6 +29,8 @@ The format is based on [Keep a Changelog 1.0.0].
 - markdown in help text fields is now parsed into HTML
 - show and hide more than one additional question at a time
 - users can be presenting with forking/branching question chains based on any one answer
+- create an extended further information text box option for radio answers
+- create extended further information text box option for checkbox answers
 
 ## [release-005] - 2021-1-19